+++ CHANGELOG.md
@@ -9,8 +9,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - ReleaseDate
 - Change default log level to `Warn`
-- Change start messages a bit to be more clear
+- Change some messages a bit to be more clear
 - Add `--print-completions` to print shell completions for various supported shells [#482](https://github.com/svenstaro/miniserve/pull/482) (thanks @rouge8)
+- Don't print some messages if not attached to an interactive terminal
+- Refuse to start if not attached to interactive terminal and no explicit path is provided
+
+  This is a security consideration as you wouldn't want to run miniserve without an explicit path
+  as a service. You could end up serving `/` or `/root` in case those working directories are set.
 
 ## [0.12.1] - 2021-03-27
 - Fix QR code not showing when using both `--random-route` and `--qrcode` [#480](https://github.com/svenstaro/miniserve/pull/480) (thanks @rouge8)